+++ src/core/component/directives/in-view/index.ts
@@ -13,7 +13,7 @@ import { ComponentDriver } from 'core/component/engines';
 
 import MutationObserverStrategy from 'core/component/directives/in-view/mutation';
 import IntersectionObserverStrategy from 'core/component/directives/in-view/intersection';
-import { DirectiveOptions, InitOptions } from 'core/component/directives/in-view/interface';
+import { DirectiveOptions, AdapteeType } from 'core/component/directives/in-view/interface';
 
 export { default as InViewAdapter } from 'core/component/directives/in-view/adapter';
 export * from 'core/component/directives/in-view/interface';
@@ -24,20 +24,28 @@ const Adaptee = getAdaptee([
 	MutationObserverStrategy
 ]);
 
+const strategyByType = {
+	mutation: MutationObserverStrategy,
+	observer: IntersectionObserverStrategy
+};
+
 /**
  * Creates a new in-view instance
+ * @param [adaptee]
  */
-export function inViewFactory(): InViewAdapter {
-	const inView = new InViewAdapter();
+export function inViewFactory(adaptee?: AdapteeType): InViewAdapter {
+	const
+		inView = new InViewAdapter(),
+		adapteeInstance = adaptee != null ? new strategyByType[adaptee]() : new Adaptee!();
 
 	if (!inView.hasAdaptee) {
-		inView.setInstance(new Adaptee!());
+		inView.setInstance(adapteeInstance);
 	}
 
 	return inView;
 }
 
-export let
+export const
 	InView: InViewAdapter = inViewFactory();
 
 if (!InView.hasAdaptee) {
@@ -50,11 +58,10 @@ ComponentDriver.directive('in-view', {
 			return;
 		}
 
-		InView.observe(el, <CanArray<InitOptions>>value);
+		InView.observe(el, value);
 	},
 
 	unbind(el: Element): void {
-		InView.stopObserve(el);
 		InView.remove(el);
 	}
 });