+++ src/core/component/directives/in-view/mutation/helpers.ts
@@ -48,8 +48,7 @@ export function isElementVisible(rect: {width: number; height: number}): boolean
 export function getElementRect(root: RootRect, el: Element): ElementRect {
 	const
 		rect = el.getBoundingClientRect(),
-		width = rect.width,
-		height = rect.height,
+		{width, height} = rect,
 		top = getOffsetTop(root, rect),
 		left = getOffsetLeft(root, rect);
 
@@ -67,7 +66,7 @@ export function getElementRect(root: RootRect, el: Element): ElementRect {
  * Returns the page root
  */
 export function getRoot(): Element {
-	return document.documentElement || document.scrollingElement || document.body;
+	return document.documentElement;
 }
 
 /**
@@ -101,6 +100,10 @@ interface RootRect {
  * @param threshold
  */
 export function isElementInView(elRect: ElementRect, rootRect: RootRect, threshold: number): boolean {
+	if (elRect.width === 0 || elRect.height === 0) {
+		return false;
+	}
+
 	const
 		isBoxInRootY = rootRect.scrollTop + rootRect.height >= elRect.top + elRect.height * threshold,
 		isBoxInRootX = rootRect.scrollLeft + rootRect.width >= elRect.left + elRect.width * threshold,